+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Change Log
     - Make status bar thicker.
     - Change default settings of items to display in toolbar/status bar.
 - Change the UI of the Unicode code point input to display the input field just above the insertion point.
+- Remove feature to import the legacy syntax style file format that was used in CotEditor 1.x. 
 - [trivial] Optimize the line number calculation in vertical text orientation.
 - [dev] Replace DifferenceKit package with native CollectionDifference.